Correciones previas merge

+++ Memoria/src/server.c
@@ -380,7 +380,7 @@ pcb_t* recibirPCB(int socket)
 	paquete->buffer = malloc(sizeof(t_buffer));
 	t_buffer* buffer = paquete->buffer;
 	//recibimos el codigo del tipo de mensaje que nos llega
-	recv(socket, &(paquete->codigo_operacion), sizeof(uint32_t), 0);
+	recv(socket, &(paquete->codigo_operacion), sizeof(uint8_t), 0);
 
 	//recibo el tamaño del paquete
 	recv(socket, &(buffer->size), sizeof(uint32_t), 0);
